warnings corrections

+++ SDK/topdown/systems/camera_system.cpp
@@ -80,7 +80,7 @@ void camera_system::process_entities(world& owner) {
 					camera.last_interpolant = camera.last_interpolant * averaging_constant + transform.pos * (1.0f - averaging_constant);
 					
 					auto interp = [](int& a, int& b, float averaging_constant){
-						a = a * averaging_constant + b * (1.0f - averaging_constant);
+						a = static_cast<int>(a * averaging_constant + b * (1.0f - averaging_constant));
 					};
 
 					interp(camera.last_ortho_interpolant.l, camera.ortho.l, averaging_constant);
